feature: akka workflow engine
